# docs/openapi/innovation-sandbox-api.yaml
openapi: 3.0.0
info:
  title: Innovation Sandbox on AWS
  description: API documentation for the Innovation Sandbox on AWS solution.
  version: 1.3.0
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "{apiBaseUrl}"
    description: Base URL used to make API requests
    variables:
      apiBaseUrl:
        default: https://example.com/api
        description: Base URL used to make API requests
...
security:
  - sigV4: []
    isbIdentity: [] # human callers (Identity Pool roles)
  - sigV4: [] # M2M callers (per-client assumed roles)
components:
  securitySchemes:
    sigV4:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: awsSigv4
      description: >-
        AWS Signature Version 4 over the `execute-api` service. Required on
        every request: the API uses IAM authorization
        (`AuthorizationType.IAM`), and there is no bearer-token or Cognito
        authorizer. The caller type is determined from the IAM principal.
    isbIdentity:
      type: apiKey
      name: x-isb-identity
      in: header
      description: >-
        Cognito ID token, verified via JWKS. Carries the claims used for RBAC.
        Required for human callers (Identity Pool roles). M2M callers (per-client
        assumed IAM roles) derive their role from the assumed-role ARN and must
        NOT send this header; a request presenting both an M2M principal and
        `x-isb-identity` is rejected with 400.
  schemas:
    ResponseBody:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - success
            - fail
            - error
      discriminator:
        propertyName: status
        mapping:
          success: "#/components/schemas/SuccessResponseBody"
          fail: "#/components/schemas/FailResponseBody"
          error: "#/components/schemas/ErrorResponseBody"
    SuccessResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - data
          properties:
            status:
              type: string
              enum:
                - success
            data:
              type: object
    FailResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - data
          properties:
            status:
              type: string
              enum:
                - fail
            data:
              type: object
    ErrorResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - message
          properties:
            status:
              type: string
              enum:
                - error
            message:
              type: string
            code:
              type: number
            data:
              type: object
...
    MetaData:
      type: object
      properties:
        createdTime:
          type: string
          format: date-time
          description: Time when the resource was created
          example: "2023-05-01T12:00:00Z"
        lastEditTime:
          type: string
          format: date-time
          description: Time when the resource was last edited
          example: "2023-05-01T12:00:00Z"
        schemaVersion:
          type: integer
          description: Version of the schema used to create the resource
          example: 1
...
    GlobalConfiguration:
      type: object
      required:
        - leases
        - cleanup
        - notification
        - maintenance
        - termsOfService
        - costReporting
        - isbManagedRegions
        - awsAccessPortalUrl
      properties:
        leases:
          type: object
          description: Lease policy configuration section
          properties:
            requireMaxBudget:
              type: boolean
            maxBudget:
              type: integer
              minimum: 0
            requireMaxDuration:
              type: boolean
            maxDurationHours:
              type: integer
              minimum: 0
            maxLeasesPerUser:
              type: integer
              minimum: 1
            ttl:
              type: integer
              minimum: 1
              description: Number of days the solution stores expired lease records before purging them
            allowUserLeaseTermination:
              type: boolean
              description: Whether users can self-terminate their own active leases
            leaseRequestWindowHours:
              type: integer
              minimum: 1
              description: Rolling window duration in hours for rate limiting lease requests
            maxLeaseRequestsPerWindow:
              type: integer
              minimum: 1
              description: Maximum number of leases a user can request within the rolling window
            leaseSharingEnabled:
              type: boolean
              description: Whether the multi-user lease sharing feature is globally enabled
            enablePrincipalSearch:
              type: boolean
              description: Whether the GET /principals/search typeahead API is available
            lastSavedBy:
              type: string
              nullable: true
              description: Email of the last admin who saved this section, or null if never saved (using defaults)
            meta:
              $ref: "#/components/schemas/MetaData"
        cleanup:
          type: object
          description: Account cleanup orchestration configuration section
          properties:
            numberOfFailedAttemptsToCancelCleanup:
              type: integer
              minimum: 1
            waitBeforeRetryFailedAttemptSeconds:
              type: integer
              minimum: 1
            numberOfSuccessfulAttemptsToFinishCleanup:
              type: integer
              minimum: 1
            waitBeforeRerunSuccessfulAttemptSeconds:
              type: integer
              minimum: 1
            validation:
              type: object
              properties:
                failureAction:
                  type: string
                  enum: [Quarantine, Warn, Silent]
                  description: Action when post-cleanup Resource Explorer validation finds remaining resources
            cooldownPeriodHours:
              type: integer
              minimum: 0
              maximum: 8640
              description: Hours to hold account in CleanUp OU after successful cleanup before returning to Available (0 = disabled, max 8640 = 360 days)
            reportRetentionDays:
              type: integer
              minimum: 14
              maximum: 3650
              description: Days to retain cleanup report records in DynamoDB
            lastSavedBy:
              type: string
              nullable: true
            meta:
              $ref: "#/components/schemas/MetaData"
        notification:
          type: object
          description: Email notification configuration section
          properties:
            emailFrom:
              type: string
              description: From address used for notification emails. Empty string disables notifications.
              example: "isb-noreply@example.com"
            lastSavedBy:
              type: string
              nullable: true
            meta:
              $ref: "#/components/schemas/MetaData"
        maintenance:
          type: object
          description: Maintenance mode configuration section
          properties:
            enabled:
              type: boolean
              description: When true, prevents creation of new leases. Fresh installs default to true.
            lastSavedBy:
              type: string
              nullable: true
            meta:
              $ref: "#/components/schemas/MetaData"
        termsOfService:
          type: object
          description: Terms of service configuration section
          properties:
            content:
              type: string
              maxLength: 10000
              description: Terms of service text displayed to users before requesting a lease
            lastSavedBy:
              type: string
              nullable: true
            meta:
              $ref: "#/components/schemas/MetaData"
        costReporting:
          type: object
          description: Cost reporting configuration section
          properties:
            costReportGroups:
              type: array
              items:
                type: string
                maxLength: 50
              maxItems: 100
              description: List of valid cost report group values for lease templates
            requireCostReportGroup:
              type: boolean
              description: Whether cost report group is required on lease templates
            lastSavedBy:
              type: string
              nullable: true
            meta:
              $ref: "#/components/schemas/MetaData"
        isbManagedRegions:
          type: array
          items:
            type: string
          minItems: 1
          description: AWS regions managed by Innovation Sandbox (read-only, resolved at deploy time)
          example: ["us-east-1", "us-west-2"]
        awsAccessPortalUrl:
          type: string
          format: uri
          description: The AWS Access Portal URL for IAM Identity Center SSO (read-only, resolved at deploy time)
...
  responses:
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
    UnauthenticatedResponse:
      description: >-
        Unauthenticated. The `x-isb-identity` ID token is missing, expired,
        invalid, or lacks required claims.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors:
                [
                  {
                    "message": "Token is missing required claims. Please re-authenticate.",
                  },
                ]
    UnauthorizedResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            message: Unauthorized
    ResourceNotFoundResponse:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
...
    ErrorResponse:
      description: Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseBody"
          example:
            status: error
            message: "An unexpected error occurred"
paths:
...
  /configurations:
    get:
      tags:
        - configurations
      summary: Get all configuration sections
      description: >
        Returns all configuration sections merged with code defaults, plus read-only
        deploy-time fields (isbManagedRegions, awsAccessPortalUrl).
        Sections never saved to DynamoDB are returned with code defaults and lastSavedBy: null.
        Accessible by Admin, Manager, and User roles.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/GlobalConfiguration"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /configurations/{section}:
    parameters:
      - name: section
        in: path
        required: true
        schema:
          type: string
          enum:
            [
              leases,
              cleanup,
              notification,
              maintenance,
              termsOfService,
              costReporting,
            ]
        description: Configuration section name. Returns 404 for unknown values.
    get:
      tags:
        - configurations
      summary: Get a single configuration section
      description: >
        Returns one configuration section with its current values (or code defaults if
        never saved) and lastSavedBy audit field.
        Returns 404 if the section name is not recognised.
        Returns 500 if the stored record fails schema validation (fails fast, unlike GET /configurations which skips invalid sections).
        Accessible by Admin, Manager, and User roles.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        type: object
                        description: Section fields plus lastSavedBy and meta
                        properties:
                          lastSavedBy:
                            type: string
                            nullable: true
                            description: Email of the last admin who saved this section, or null if never saved
                          meta:
                            $ref: "#/components/schemas/MetaData"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags:
        - configurations
      summary: Save a configuration section (full replacement)
      description: >
        Replaces an entire configuration section. All section fields are required.
        Uses optimistic concurrency: provide meta.lastEditTime from a prior GET to
        detect concurrent edits (returns 409 on conflict). Omit meta.lastEditTime on
        the first save for a section (uses attribute_not_exists condition).

        M2M callers are rejected with 403 — audit trail requires a human email identity.
        Admin role only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                All fields for the section (required, no defaults applied server-side).
                Plus optional audit envelope for optimistic concurrency.
              properties:
                meta:
                  type: object
                  description: Optional optimistic concurrency envelope
                  properties:
                    lastEditTime:
                      type: string
                      format: date-time
                      description: Value from the prior GET response. Omit on first save.
              additionalProperties: true
              example:
                requireMaxBudget: true
                maxBudget: 100
                requireMaxDuration: true
                maxDurationHours: 168
                maxLeasesPerUser: 3
                ttl: 30
                allowUserLeaseTermination: true
                leaseRequestWindowHours: 168
                maxLeaseRequestsPerWindow: 10
                leaseSharingEnabled: false
                enablePrincipalSearch: true
                meta:
                  lastEditTime: "2026-04-04T12:30:00.000Z"
      responses:
        "200":
          description: Section saved successfully. Returns updated section with new lastEditTime and lastSavedBy.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        type: object
                        description: Updated section fields plus lastSavedBy and updated meta
                        properties:
                          lastSavedBy:
                            type: string
                            description: Email of the admin who saved this section
                          meta:
                            $ref: "#/components/schemas/MetaData"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          description: Section not found (unknown section name)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponseBody"
        "409":
          description: Optimistic concurrency conflict — section was modified by another admin. Reload and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponseBody"
              example:
                status: fail
                data:
                  errors:
                    - message: "Configuration was modified by another administrator. Reload to see the latest values."
        "500":
          $ref: "#/components/responses/ErrorResponse"

  # Principals endpoints
...